==> 10_planned_analyses/COMP_ECA/scripts/4_pcaPlots.Rmd <==
---
title: "4_pcaPlots"
author: "Paul A. Bloom"
date: "September 10, 2019"
output:
  html_document:
    df_print: paged
  png_document: default
---

# notes 
PCA thing said to look at 2 dimensions and 3 factors 


```{r, include = F, warnings = F, echo=FALSE, message = F}
source("../../../0_R_analysis_setup_file.R")
```

# get data organized
## Load in model and pacct raw data (clean same way as orig bootstrap PCA script)
```{r effortdata}
load('../PCA_output/pcaFullSampleModel.rda') # full sample model output

load("../data/pacct_effort_w_labels.Rdata")
```

## Load in Raw CBCL through master
```{r master}
master <- read_excel("../../../../subj_demo/masters/MASTER_20191118_updated_MICM.xlsx") 
master = master %>% 
  dplyr::select(IDENT_SUBID, CBCL_4_INTERNPROB_T, CBCL_4_EXTERNPROB_T, CBCL_4_TOTPROB_T) 
```


## combine PCA results, original labeled data, and CBCL 
```{r combine}
pcaScoresMat = pcaFull$scores

# Join up PCA scores matrix with PACCT group labels, CBCL scores
pacctPCAScores = cbind(pacct_effort_w_labels, pcaScoresMat) %>%
  left_join(., master) %>%
  mutate(Sex = ifelse(DEM_3_GENDER_CHILD == '0', 'male', 'female'), 
         GROUP = as.factor(GROUP))
#pacctPCAScores$GROUP
summary(as.factor(pacct_effort_w_labels$GROUP))
```


# Plots of variable loadings onto PCs

## Prep bootstrap matching for plots
```{r prep}
load('../PCA_output/matched_pca.rda')
load('../PCA_output/pcaFullSample.rda')

fullSampLoadings = data.frame(pcaFullSample[[2]])
fullSampLoadings$vars = row.names(fullSampLoadings)


fullSampLoadingsLong = fullSampLoadings %>%
  tidyr::gather(., key = 'dimension', value = 'loading', dim.1:dim.3) %>%
  mutate(., dimension = as.numeric(gsub('dim.', '', dimension)))

# Loop through to put pca_matching list object into big dataframe
for (i in 1:length(pca_matching_loadings)){
  tmp = data.frame(pca_matching_loadings[[i]])
  names(tmp) = paste0('dim', 1:ncol(tmp))
  tmp$vars = row.names(tmp)
  tmp$iter = i
  if (i == 1){
    bootMatch = tmp 
  }
  else{
    bootMatch = rbind(bootMatch, tmp)
  }
}

# put bootMatch in long form, then summarize
bootMatchSummary = bootMatch %>%
  tidyr::gather(., key = 'dimension', value = 'loading', dim1:dim10) %>%
  dplyr::mutate(., dimension = as.numeric(gsub('dim', '', dimension))) %>%
  dplyr::group_by(vars, dimension) %>%
  dplyr::summarise(meanLoad = mean(loading),
                   sdLoad = sd(loading),
                   coefVar = meanLoad/sdLoad,
            lwr95 = quantile(loading, probs = .025),
            upr95 = quantile(loading, probs = .975),
            lwr80 = quantile(loading, probs = .1),
            upr80 = quantile(loading, probs = .9))

```

## plot the bootstrapped absolute value loadings
```{r plotboot}
loadingBootMatchPlot = 
  dplyr::filter(bootMatchSummary, dimension <= 3) %>%
  ggplot(., aes(x = vars, y = meanLoad)) +
  geom_hline(yintercept =0, color = 'red', lwd = .5) +
  geom_errorbar(aes(ymin = lwr95, ymax = upr95), width = 0, alpha = .5) +
  geom_errorbar(aes(ymin = lwr80, ymax = upr80), width = 0, lwd = 1) +
  geom_point() +
  geom_point(data = fullSampLoadingsLong, aes(x = vars, y = loading), color = 'red', size = .5) +
  facet_grid(~dimension) +
  coord_flip() +
  theme_bw() +
  labs(title = 'Boostrap (w/matching) Distributions of Variable loadings on PC 1-3\nBlue points are loadings from PCA fit to full sample', y = 'Abs. Value Loading')

loadingBootMatchPlot
ggsave(loadingBootMatchPlot, height = 8, width = 12, file = '../figures/PCA/pcaLoadingBootMatchPlot.png')

# Also save out bootstrap matching loadings for the 2-way clustering
bootMatchLoadingsFor2WayClust = bootMatchSummary %>%
  dplyr::select(., vars, dimension, meanLoad) %>%
  tidyr::spread(key = dimension, value = meanLoad)

save(bootMatchLoadingsFor2WayClust, file = '../PCA_output/bootstrapLoadingMeans.rda')

```



## Run permutation test to threshold raw loadings (how stable?)
```{r permutationtest}
# get the original PCA input data.
load("../data/cleaned_ELV_for_dim_reduction.Rdata")

# make a null distribution to compare our correrlatins with
nPerms = 1000
corVec = rep(NA, nPerms) 
for (jj in 1:nPerms){
  var1 = rnorm(n = nrow(pacctNoLabelsCleaned))
  var2 = rnorm(n = nrow(pacctNoLabelsCleaned))
  corVec[jj] = cor(var1, var2)
}
hist(abs(corVec))

# make the boundary the 99% percentile?
boundary = quantile(abs(corVec), prob = .99)

# load the rraw loadings 
load('../PCA_output/pcaFullLoadingsNotAbsVal.rda')
regularLoadings = mutate(regularLoadings, dimension = parse_number(dimensions)) %>%
  select(-dimensions) %>%
  left_join(., bootMatchSummary, by = c('vars', 'dimension')) %>%
  # indicate which loadings are stable
  mutate(., stable = ifelse(meanLoad > boundary, 'Stable', 'Unstable'))
```

## plot thresholded non-absolute value loadings (raw)
color-coded with how significant/stable the loadings are. 
this is quite different than ECA-only analysis with CMI...
PC1 = cumulative adversity, stable on all items. 
PC2 =  positive on neglect items, negative on many threat items.
PC3 = positive on physical abuse, negative on sexual abuse. 

```{r plotloadings}
loadingsNoAbsValPlot = ggplot(regularLoadings, aes(x = vars, y = loading, color = stable)) +
  geom_hline(yintercept = 0) +
  geom_point() + 
  facet_grid(~dimension) + 
  coord_flip() + 
  theme_bw() 

loadingsNoAbsValPlot

ggsave(loadingsNoAbsValPlot, height = 8, width = 12, file = '../figures/PCA/pcaLoadingFullNotAbsoluteValued.png')
```

# Subject Scores (predictions) of PCs
## get bootstrapped scores into better format
```{r getscores}
# Loop through to put subject-level scores all in one data frame
for (i in 1:length(pca_matching_preds)){
  
  test <- data.frame(pca_matching_preds[[1]])
  tmp = data.frame(pca_matching_preds[[i]])
  names(tmp) = paste0('dimBoot', 1:ncol(tmp))
  tmp$subject = floor(as.numeric(row.names(tmp)))
  tmp$iter = i
  if (i == 1){
    bootMatchPreds = tmp 
  }
  else{
    bootMatchPreds = rbind(bootMatchPreds, tmp)
  }
}



# load in key and orig scores
load('../PCA_output/pcaFullScoresKey.rda')

pcaFullScores$subject = as.numeric(pcaFullScores$subject)
max(bootMatchPreds$subject)
max(pcaFullScores$subject)

# Generate summary dataframe
bootMatchPredSummary = bootMatchPreds %>%
  tidyr::gather(., key = 'dimension', value = 'prediction', dimBoot1:dimBoot3) %>%
  group_by(dimension, subject) %>%
  summarise(meanPred = mean(prediction)) %>%
  tidyr::spread(key = dimension, value = meanPred) %>%
  left_join(., pcaFullScores) %>%
  left_join(., pacct_effort_w_labels) %>%
  left_join(., master) %>%
  mutate(Sex = ifelse(DEM_3_GENDER_CHILD == '0', 'male', 'female'), 
         GROUP = as.factor(GROUP))
```

## Correlate full sample vs. bootstrapped subject scores, for each dimension
not correlated!!! 
```{r plotcorr}

# corrplot only first 5 dimensions
bootMatchPredsFirst5Cor = dplyr::select(bootMatchPredSummary, contains('boot'), dim.1:dim.3) %>%
  cor() %>%
  corrplot(type = 'lower')

# write out subject-level loadings to csv, for comparison with factor analysis
bootstrapPCASubjectLoadings = bootMatchPredSummary %>%
  dplyr::select(contains('Boot'), GROUP, Sex, IDENT_SUBID)

write.csv(bootstrapPCASubjectLoadings, file = '../PCA_output/bootstrapPCASubjectLoadings.csv', row.names = FALSE)

```

## Another way: plot Subject-level predictions from full sample vs. bootstrap mean

```{r plotsubscorr}
  
comparingSubjectScoresDim1 <- ggplot(bootMatchPredSummary, aes(x = dimBoot1, y = dim.1)) +
  geom_point(aes(color = GROUP)) +
  geom_abline(slope = 1, intercept = 0) +
  theme_bw() 

comparingSubjectScoresDim2 <- ggplot(bootMatchPredSummary, aes(x = dimBoot2, y = dim.2)) +
  geom_point(aes(color = GROUP)) +
  geom_abline(slope = 1, intercept = 0) +
  theme_bw()

comparingSubjectScoresDim1
comparingSubjectScoresDim2

ggsave(comparingSubjectScoresDim1, file = "../figures/PCA/comparingSubjectScoresDim1.png")
ggsave(comparingSubjectScoresDim2, file = "../figures/PCA/comparingSubjectScoresDim2.png")

```

## Plot full Sample subject scores by group
these make sense.
```{r plotsubgroups}

fullSampGroupPC1_2 = ggplot(pacctPCAScores, aes(x = `dim 1`, y = `dim 2`)) +
  geom_point(aes(color = GROUP, pch = Sex), size = 2, alpha = 0.7) +
  theme_bw() 

fullSampGroupPC1_3 = ggplot(pacctPCAScores, aes(x = `dim 1`, y = `dim 3`)) +
  geom_point(aes(color = GROUP, pch = Sex), size = 2, alpha = 0.7) +
  theme_bw() 


fullSampGroupPC2_3 = ggplot(pacctPCAScores, aes(x = `dim 2`, y = `dim 3`)) +
  geom_point(aes(color = GROUP, pch = Sex), size = 2, alpha = 0.7) +
  theme_bw() 

fullSampGroupPC1_2 
fullSampGroupPC1_3
fullSampGroupPC2_3

ggsave(fullSampGroupPC1_2, file = '../figures/PCA/pcProjection_1_2_group.png')
ggsave(fullSampGroupPC1_3, file = '../figures/PCA/pcProjection_1_3_group.png')
ggsave(fullSampGroupPC2_3, file = '../figures/PCA/pcProjection_2_3_group.png')

```

## plot bootstrapped subject scores by group 
makes sense now
```{r plotbootgroups}
bootGroupPC1_2 = ggplot(bootMatchPredSummary, aes(x = dimBoot1, y = dimBoot2)) +
  geom_point(aes(color = GROUP, pch = Sex), size = 2, alpha = 0.7) +
  theme_bw() 

bootGroupPC1_3 = ggplot(bootMatchPredSummary, aes(x = dimBoot1, y = dimBoot3)) +
  geom_point(aes(color = GROUP, pch = Sex), size = 2, alpha = 0.7) +
  theme_bw()

bootGroupPC1_2 
bootGroupPC1_3 


ggsave(bootGroupPC1_2, file = '../figures/PCA/pcProjectionBootstrapMean_1_2_group.png')
ggsave(bootGroupPC1_3, file = '../figures/PCA/pcProjectionBoostrapMean_1_3_group.png')
```



## Scatter plot of full sample PCs with CBCL
this looks reasonable!
```{r plotcbcl}
pacctPCAScoresLong = pacctPCAScores %>%
  tidyr::gather(key = 'dimension', value = 'loading', contains('dim')) %>%
  mutate(dimNum = parse_number(dimension))

cbclPC1 = ggplot(filter(pacctPCAScoresLong, dimNum < 2), 
                 aes(x = loading, y = CBCL_4_INTERNPROB_T)) +
  geom_point() + 
  stat_smooth(method = 'lm') +
  theme_bw() +
  labs(title = 'CBCL scores as a function of PC1')

cbclPC1

ggsave(cbclPC1, file = '../figures/PCA/pcdim1_fullsample_cbcl.png', height = 6, width = 8)

```

## Scatter plot of bootstrapepd PCs with CBCL 
```{r plotbootcbcl}
bootMatchPredSummaryLong <- bootMatchPredSummary %>%
  select(-starts_with('dim.')) %>%
  tidyr::gather(key = 'dimension', value = 'loading', contains('dimBoot')) %>%
  mutate(dimNum = parse_number(dimension))

cbclPC1 = ggplot(filter(bootMatchPredSummaryLong, dimNum < 2), 
                 aes(x = loading, y = CBCL_4_INTERNPROB_T)) +
  geom_point() + 
  stat_smooth(method = 'lm') +
  theme_bw() +
  labs(title = 'CBCL scores as a function of PC1')

cbclPC1

ggsave(cbclPC1, file = '../figures/PCA/pcdim1_boot_cbcl.png', height = 6, width = 8)

```

# Scree plots with confidence from bootstrapped models

## get data into better format 
```{r getscree}
for (i in 1:length(pca_matching_eig)){
  tmp = data.frame(pca_matching_eig[[i]])
  tmp$dimension = as.numeric(gsub('dim ', '', row.names(tmp)))
  tmp$iter = i
  if (i == 1){
    bootMatchEig = tmp 
  }
  else{
    bootMatchEig = rbind(bootMatchEig, tmp)
  }
}

bootMatchEigSummary = bootMatchEig %>%
  group_by(dimension) %>%
  summarise(meanEig= mean(Eigenvalue),
            eigLwr95 = quantile(Eigenvalue, probs = .025),
            eigUpr95 = quantile(Eigenvalue, probs = .975),
            eigLwr80 = quantile(Eigenvalue, probs = .1),
            eigUpr80 = quantile(Eigenvalue, probs = .9),
            meanProp= mean(Proportion),
            propLwr95 = quantile(Proportion, probs = .025),
            propUpr95 = quantile(Proportion, probs = .975),
            propLwr80 = quantile(Proportion, probs = .1),
            propUpr80 = quantile(Proportion, probs = .9),
            meanCumulative= mean(Cumulative),
            cumulativeLwr95 = quantile(Cumulative, probs = .025),
            cumulativeUpr95 = quantile(Cumulative, probs = .975),
            cumulativeLwr80 = quantile(Cumulative, probs = .1),
            cumulativeUpr80 = quantile(Cumulative, probs = .9))
```

## make plots
```{r plotscree}
# Eigenvalue plot
screeEig = ggplot(data = bootMatchEigSummary, aes(x = dimension, y = meanEig)) +
  geom_errorbar(aes(ymin = eigLwr95, ymax = eigUpr95), alpha = .7, width = .05) +
  geom_errorbar(aes(ymin = eigLwr80, ymax = eigUpr80), width = 0, lwd = 1) +
  geom_point(size = 2, color = 'purple') +
  scale_x_continuous(labels = 1:10, breaks = 1:10) +
  theme_bw() +
  labs(x = 'PC', y = 'Eigenvalue', title = 'Scree Plot From Matching Boostrap (Eigenvalues)') +
  theme(panel.grid.minor = element_blank(), panel.grid.major.x = element_blank())

# Proportion variance explained plot
screeProp = ggplot(data = bootMatchEigSummary, aes(x = dimension, y = meanProp)) +
  geom_errorbar(aes(ymin = propLwr95, ymax = propUpr95), alpha = .7, width = .05) +
  geom_errorbar(aes(ymin = propLwr80, ymax = propUpr80), width = 0, lwd = 1) +
  geom_point(size = 2, color = 'purple') +
  scale_x_continuous(labels = 1:10, breaks = 1:10) +
  theme_bw() +
  labs(x = 'PC', y = 'Proportion Variance Explained', title = 'Scree Plot From Matching Boostrap (Proportion Variance)') +
  theme(panel.grid.minor = element_blank(), panel.grid.major.x = element_blank())

# Cumulative proportion variance explained plot
screeCumulative = ggplot(data = bootMatchEigSummary, aes(x = dimension, y = meanCumulative)) +
  geom_line() + 
  geom_errorbar(aes(ymin = cumulativeLwr95, ymax = cumulativeUpr95), alpha = .7, width = .05) +
  geom_errorbar(aes(ymin = cumulativeLwr80, ymax = cumulativeUpr80), width = 0, lwd = 1) +
  geom_point(size = 2, color = 'purple') +
  theme_bw() +
  scale_x_continuous(labels = 1:10, breaks = 1:10) +
  labs(x = 'PC', y = 'Cumulative Proportion Variance Explained', title = 'Scree Plot From Matching Boostrap (Cumulative Variance)') +
  ylim(0,100) +
  theme(panel.grid.minor = element_blank(), panel.grid.major.x = element_blank())

screeEig
screeProp
screeCumulative

ggsave(screeEig, file = '../figures/PCA/screeEigenvaluesPCA_MatchBoostrap.png')
ggsave(screeProp, file = '../figures/PCA/screePropVariancePCA_MatchBoostrap.png')
ggsave(screeCumulative, file = '../figures/PCA/screeCumulativeVariancePCA_MatchBoostrap.png')

```


## save Full Sample & bootstrap subject scores  for future use!
```{r savescree}
pacctPCAScores2 <- pacctPCAScores %>%
  select(SUBJECTID, GROUP, Sex, 'dim 1') %>%
  rename(SamplePC1 = 'dim 1')
nrow(pacctPCAScores2)
bootMatchPredSummary2 <- bootMatchPredSummary %>%
  rename(BootstrapPC1 = 'dimBoot1', 
         BootstrapPC2 = 'dimBoot2') %>%
  select(SUBJECTID, BootstrapPC1, BootstrapPC2, CBCL_4_INTERNPROB_T, CBCL_4_EXTERNPROB_T, CBCL_4_TOTPROB_T)
nrow(bootMatchPredSummary2)
pacctPCAscores <- left_join(pacctPCAScores2, bootMatchPredSummary2)

save(pacctPCAscores, file= "../PCA_output/pcaFullSampleScoresWithLabels.rda")

```
